requirements.txt
numpy
Django==1.8.4
python-dateutil
matplotlib==1.4.3
netCDF4>=1.0.2
rtree==0.8.2
jsonfield==1.0.3
pytz
djangorestframework==3.2.4
pyugrid
pysgrid>=0.1.0
pyaxiom>=0.0.9
django-grappelli==2.7.1
django-typed-models==0.5.0
pyproj
dj-static==0.0.6
pandas
django-autoslug==1.9.3
isodate
utide

celery==3.1.20
hiredis
django-redis-cache==1.6.0
django-celery==3.1.17
